feature(pivot): preserving expanded columns when changing sort
