+++ src/exampleTypes.js
@@ -1,6 +1,6 @@
 import _ from 'lodash/fp'
 import F from 'futil'
-import { deepMultiTransformOn } from './util/futil'
+import pivot from './pivot'
 
 let validateValues = ({ value, values = [] }) => value || values.length
 let validateValueExistence = _.flow(_.get('value'), _.negate(_.isNil))
@@ -18,20 +18,6 @@ let twoLevelMatch = {
     key_value: '',
   },
 }
-
-let getKey = x => x.keyAsString || x.key
-
-// Similar to Tree.lookup but path is a drilldown which uses keyAsString or key
-let resultsForDrilldown = (type, path, results) => {
-  if (_.isEmpty(path) || !results) return results
-
-  let key = _.first(path)
-  let groups = _.get(type, results)
-  let match = _.find(node => getKey(node) === key, groups)
-  return resultsForDrilldown(type, path.slice(1), match)
-}
-
-let someNotEmpty = _.some(_.negate(_.isEmpty))
 
 export default F.stampKey('type', {
   facet: {
@@ -279,199 +265,7 @@ export default F.stampKey('type', {
       field: null,
     },
   },
-  pivot: {
-    validate: context =>
-      _.every(
-        ({ type, ranges, percents }) =>
-          (type !== 'numberRanges' && type !== 'percentiles') ||
-          (type === 'numberRanges' && ranges.length > 0) ||
-          (type === 'percentiles' && percents.length > 0),
-        _.concat(context.columns, context.rows)
-      ),
-    reactors: {
-      columns: 'self',
-      rows: 'self',
-      values: 'self',
-      pagination: 'self',
-      filters: 'others',
-      sort: 'self',
-    },
-    defaults: {
-      columns: [],
-      rows: [],
-      values: [],
-      filters: [],
-      sort: {},
-      pagination: {
-        columns: {
-          drilldown: null,
-          skip: [],
-          expanded: [],
-        },
-        rows: {
-          drilldown: null,
-          skip: [],
-          expanded: [],
-        },
-      },
-      showCounts: false,
-      context: {
-        results: [],
-      },
-    },
-    onDispatch(event, extend) {
-      let { type, node, previous, value } = event
-
-      if (type === 'mutate') {
-        let prevRowDrill = _.get('pagination.rows.drilldown', previous)
-        let prevColumnDrill = _.get('pagination.columns.drilldown', previous)
-
-        // If mutation is a pagination
-        if (
-          _.has('pagination.columns', value) ||
-          _.has('pagination.rows', value)
-        ) {
-          let isRowPagination = _.has('pagination.rows', value)
-
-          // Getting previous drilldown/skip values and adding include
-          let getPrevPage = type => {
-            let page = _.flow(
-              _.get(['pagination', type]),
-              _.pick(['drilldown', 'skip'])
-            )(previous)
-
-            if (page) {
-              // adding root level include for the first page only if expanded is empty yet
-              // allows us to remember values of existing root level rows/columns
-              if (
-                _.isEmpty(page.drilldown) &&
-                _.isEmpty(page.skip) &&
-                !_.isEmpty(_.get(['pagination', type, 'expanded'], previous))
-              )
-                return false
-
-              // getting existing results for the drilldown
-              let results = resultsForDrilldown(
-                type,
-                page.drilldown,
-                _.get('context.results', node)
-              )
-
-              // getting existing row/column values without skipped ones
-              page.include = _.flow(
-                _.get(type),
-                _.map(getKey),
-                _.without(page.skip)
-              )(results)
-            }
-
-            return page
-          }
-
-          // Previous expanded values with previous drilldown added
-          let getExpanded = type =>
-            _.compact([
-              ..._.getOr([], ['pagination', type, 'expanded'], previous),
-              getPrevPage(type),
-            ])
-
-          // Preserving previous pagination entries in the expanded prop
-          extend(node, {
-            pagination: {
-              columns: {
-                drilldown: prevColumnDrill && [],
-                skip: [],
-                ...(!isRowPagination && value.pagination.columns),
-                expanded: getExpanded('columns'),
-              },
-              rows: {
-                drilldown: prevRowDrill && [],
-                skip: [],
-                ...(isRowPagination && value.pagination.rows),
-                expanded: getExpanded('rows'),
-              },
-            },
-          })
-          // If drilldown mode is enabled for columns or rows
-        } else if (prevColumnDrill || prevRowDrill)
-          // Resetting the pagination when the pivot node is changed
-          // allows to return expected root results instead of nested drilldown
-          // EX: changing the columns or rows config was not returning the new results
-          extend(node, {
-            pagination: {
-              columns: {
-                drilldown: prevColumnDrill && [],
-                skip: [],
-                expanded: [],
-              },
-              rows: {
-                drilldown: prevRowDrill && [],
-                skip: [],
-                expanded: [],
-              },
-            },
-          })
-      }
-    },
-    // Resetting the pagination when the tree is changed
-    // allows to return expected root results instead of nested drilldown
-    // EX: criteria filters didn't work properly when drilldown was applied
-    onUpdateByOthers(node, extend) {
-      let rowDrill = _.get('pagination.rows.drilldown', node)
-      let columnDrill = _.get('pagination.columns.drilldown', node)
-      if (columnDrill || rowDrill)
-        extend(node, {
-          pagination: {
-            columns: {
-              drilldown: columnDrill && [],
-              skip: [],
-              expanded: [],
-            },
-            rows: {
-              drilldown: rowDrill && [],
-              skip: [],
-              expanded: [],
-            },
-          },
-        })
-    },
-    shouldMergeResponse: node =>
-      // checking for presence of drilldown, skip, expanded in pagination
-      someNotEmpty(
-        _.map(_.get(_, node), [
-          'pagination.columns.drilldown',
-          'pagination.columns.skip',
-          'pagination.columns.expanded',
-          'pagination.rows.drilldown',
-          'pagination.rows.skip',
-          'pagination.rows.expanded',
-        ])
-      ),
-    mergeResponse(node, response, extend, snapshot) {
-      // Convert response rows and columns to objects for easy merges
-      let groupsToObjects = deepMultiTransformOn(
-        ['rows', 'columns'],
-        groupsToObjects => _.flow(_.map(groupsToObjects), _.keyBy('key'))
-      )
-      // Convert rows and columns back to arrays
-      let groupsToArrays = deepMultiTransformOn(
-        ['rows', 'columns'],
-        groupsToArrays => _.flow(_.values, _.map(groupsToArrays))
-      )
-
-      // `snapshot` here is to solve a mobx issue
-      let nodeGroups = groupsToObjects(snapshot(node.context.results))
-      let responseGroups = groupsToObjects(response.context.results)
-
-      // Easy merge now that we can merge by row key
-      let results = F.mergeAllArrays([nodeGroups, responseGroups])
-
-      let context = { results: groupsToArrays(results) }
-
-      // Write on the node
-      extend(node, { context })
-    },
-  },
+  pivot,
   esTwoLevelAggregation: {
     validate: context =>
       context.key_field &&
